push more in sh

+++ tests/unittest/allcover.sh
@@ -1,21 +1,26 @@
-export GOROOT=/home/runner/work/hera2flow/hera2flow/go
-mkdir -p /home/runner/go/src/github.com/paypal
-ln -s /home/runner/work/hera2flow/hera2flow /home/runner/go/src/github.com/paypal/hera
+wget -nv https://go.dev/dl/go1.20.7.linux-amd64.tar.gz
+echo f0a87f1bcae91c4b69f8dc2bc6d7e6bfcd7524fceec130af525058c0c17b1b44  go1.20.7.linux-amd64.tar.gz | sha256sum -c
+if [ x$? != x ]
+then
+    echo bad sha256sum
+    exit 1
+fi
+
+export GOROOT=$PWD/go
 export GOPATH=/home/runner/go
+mkdir -p $GOPATH/src/github.com/paypal
+ln -s $PWD $GOPATH/src/github.com/paypal/hera
 
 rm -rf $GOPATH/allcover{,2}
 mkdir $GOPATH/allcover
 
 $GOROOT/bin/go install -cover github.com/paypal/hera/worker/mysqlworker
-ls -l $GOPATH/bin
-pwd
 
 overall=0
 for d in `ls -F tests/unittest | grep /$ | sed -e "s,/,," | egrep -v '(mysql_recycle|log_checker_initdb|testutil|rac_maint|mysql_direct|failover)'`
 do 
     echo ==== $d
-    pushd tests/unittest/$d 
-    #cp /home/runner/go/bin/mysqlworker .
+    cd tests/unittest/$d 
     cp $GOPATH/bin/mysqlworker .
     rm -f *.log 
 
@@ -32,8 +37,10 @@ do
     mv $GOPATH/allcover{2,}
 
     rm -f *.log 
-    popd
+    cd ../../..
 done
 $GOROOT/bin/go tool covdata func -i=$GOPATH/allcover
 $GOROOT/bin/go tool covdata percent -i=$GOPATH/allcover
+$GOROOT/bin/go tool covdata textfmt -i=$GOPATH/allcover -o $GOPATH/allcover.out
+$GOROOT/bin/go tool cover -html=$GOPATH/allcover.out -o $GOPATH/allcover.htm
 exit $overall
